simplify calls to forRange2D
