+++ index.mjs
@@ -4,9 +4,9 @@ import { solve } from "./lib/solver";
 const b = new Board();
 fillBoard(b,
   "956      ",
-  "         ",
+  "        5",
   " 8  463 7",
-  "6  28    ",
+  "6  28   1",
   "5  7  48 ",
   "2        ",
   "  5 7  4 ",
