+++ lib/traverse.mjs
@@ -36,8 +36,20 @@ export function forEveryCell(board, fn) {
   return forRange2D(0, 9, 1, 0, 9, 1, (x, y) => fn(board.getCell(x, y), x, y));
 }
 
-export function forHouse(board, xLeft, yTop, fn) {
-  return forRange2D(xLeft, xLeft + 3, 1, yTop, yTop + 3, 1, (x, y) =>
+export function forHouse(board, houseX, houseY, fn) {
+  const x1 = houseX * 3
+  const y1 = houseY * 3
+  const x2 = x1 + 3
+  const y2 = y1 + 3
+  return forRange2D(x1, x2, 1, y1, y2, 1, (x, y) =>
     fn(board.getCell(x, y), x, y)
   );
+}
+
+export function findFirstEmptyCell(board) {
+  return forEveryCell(board, (val, x, y) => (val === 0 ? [x, y] : null));
+}
+
+function isAllFull(board) {
+  return !!findFirstEmptyCell(board);
 }
